+++ src/cmdline/mod.rs
@@ -18,15 +18,15 @@
 use clap::{AppSettings, Parser};
 use reqwest::Url;
 
+#[cfg(feature = "docgen")]
+mod doc;
 mod install;
-#[cfg(feature = "mangen")]
-mod man;
 mod serializer;
 mod types;
 
+#[cfg(feature = "docgen")]
+pub use self::doc::*;
 pub use self::install::InstallConfig;
-#[cfg(feature = "mangen")]
-pub use self::man::*;
 pub use self::types::*;
 
 // Args are listed in --help in the order declared in these structs/enums.
@@ -167,8 +167,11 @@ pub enum PackCmd {
     /// Pack a minimal ISO into a CoreOS live ISO image
     MinimalIso(PackMinimalIsoConfig),
     /// Generate man pages for coreos-installer
-    #[cfg(feature = "mangen")]
+    #[cfg(feature = "docgen")]
     Man(PackManConfig),
+    /// Generate example config file for install subcommand
+    #[cfg(feature = "docgen")]
+    ExampleConfig(PackExampleConfigConfig),
 }
 
 #[derive(Debug, Parser)]
@@ -692,13 +695,17 @@ pub struct DevExtractInitrdConfig {
     pub filter: Vec<String>,
 }
 
-#[cfg(feature = "mangen")]
+#[cfg(feature = "docgen")]
 #[derive(Debug, Parser)]
 pub struct PackManConfig {
     /// Output directory
     #[clap(short = 'C', long, value_name = "path", default_value = ".")]
     pub directory: String,
 }
+
+#[cfg(feature = "docgen")]
+#[derive(Debug, Parser)]
+pub struct PackExampleConfigConfig {}
 
 #[cfg(test)]
 mod test {