+++ docs/release-notes.md
@@ -31,6 +31,7 @@ Packaging changes:
 - Add dependency on `zstd` crate and `libzstd` shared library
 - Support `serde_yaml` 0.9
 - Remove non-Linux dependencies from vendor archive
+- Install example installer config file in `/usr/share/coreos-installer`
 
 
 ## coreos-installer 0.15.0 (2022-06-17)